// packages/flare-core-typescript-sdk/src/model/message_build_op.ts
/** GENERATED. Do not edit by hand. */
/** Dispatch op passed to flare_message_build_json. */
export enum MessageBuildOp {
  CreateText = "create_text",
  CreateQuote = "create_quote",
  CreateThreadReply = "create_thread_reply",
  CreateForward = "create_forward",
  CreateImage = "create_image",
  CreateImageGroup = "create_image_group",
  CreateVideo = "create_video",
  CreateAudio = "create_audio",
  CreateFile = "create_file",
  CreateEmoji = "create_emoji",
  CreateLocation = "create_location",
  CreateSticker = "create_sticker",
  CreateLinkCard = "create_link_card",
  CreateCard = "create_card",
  CreateMiniProgram = "create_mini_program",
  CreateRichDoc = "create_rich_doc",
  CreateSystem = "create_system",
  CreateNotification = "create_notification",
  CreateVote = "create_vote",
  CreateTask = "create_task",
  CreateSchedule = "create_schedule",
  CreateAnnouncement = "create_announcement",
  CreateCustom = "create_custom",
  CreatePlaceholder = "create_placeholder",
  CreateWithContent = "create_with_content",
}